+++ cli/vendor.ts
@@ -1,40 +1,34 @@
-import { brightGreen } from "../deps/colors.ts";
 import {
   getConfigFile,
-  getDenoConfig,
-  getImportMap,
-  loadImportMap,
-  toUrl,
+  readDenoConfig,
+  writeDenoConfig,
 } from "../core/utils.ts";
 import { join } from "../deps/path.ts";
-import { promptConfigUpdate } from "./utils.ts";
 
 interface Options {
   output: string;
-  root: string;
   config?: string;
   remove?: boolean;
 }
 
-export default function ({ output, root, config, remove }: Options) {
-  return vendor(root, output, config, remove);
+export default function ({ output, config, remove }: Options) {
+  return vendor(output, config, remove);
 }
 
 /** Upgrade the Lume installation to the latest version */
 export async function vendor(
-  root: string,
   output: string,
   config?: string,
   remove?: boolean,
 ) {
-  await removeVendor(root, output);
+  await removeVendor(output);
 
   if (remove) {
     return;
   }
 
-  const configFile = await getConfigFile(root, config);
-  const vendor = join(root, output);
+  const configFile = await getConfigFile(Deno.cwd(), config);
+  const oldDenoConfig = await readDenoConfig();
 
   // Run deno vendor
   const specifiers: string[] = [
@@ -45,40 +39,25 @@ export async function vendor(
 
   await run(output, specifiers);
 
-  // Write the import map
-  const importMapFile = join(vendor, "import_map.json");
-  const denoConfigInfo = await getDenoConfig();
-  const denoConfig = denoConfigInfo?.config || {};
-  const currentMap = await getImportMap(denoConfig.importMap);
-  const vendorMap = await loadImportMap(await toUrl(importMapFile));
+  // Fix the import map
+  if (oldDenoConfig?.importMap) {
+    const denoConfig = await readDenoConfig();
 
-  for (let [specifier, path] of Object.entries(currentMap.imports)) {
-    if (vendorMap.imports[specifier]) {
-      continue;
+    if (denoConfig?.importMap) {
+      for (
+        let [specifier, path] of Object.entries(oldDenoConfig.importMap.imports)
+      ) {
+        if (denoConfig.importMap.imports[specifier]) {
+          continue;
+        }
+        if (path.startsWith(".")) {
+          path = join("./", output, path);
+        }
+        denoConfig.importMap.imports[specifier] = path;
+      }
+
+      await writeDenoConfig(denoConfig);
     }
-    if (path.startsWith(".")) {
-      path = join("./", output, path);
-    }
-    vendorMap.imports[specifier] = path;
-  }
-  await Deno.writeTextFile(
-    importMapFile,
-    JSON.stringify(vendorMap, null, 2) + "\n",
-  );
-
-  // Update deno.json file
-  denoConfig.importMap = importMapFile;
-  if (denoConfigInfo?.file === "deno.jsonc") {
-    promptConfigUpdate({ importMap: denoConfig.importMap });
-  } else {
-    await Deno.writeTextFile(
-      join(root, "deno.json"),
-      JSON.stringify(denoConfig, null, 2) + "\n",
-    );
-
-    console.log(brightGreen("Lume vendored to:"), output);
-    console.log(brightGreen("Deno configuration file saved:"), "deno.json");
-    console.log(brightGreen("Import map file saved:"), denoConfig.importMap);
   }
 }
 
@@ -100,28 +79,20 @@ async function run(output: string, urls: string[]) {
   }
 }
 
-async function removeVendor(root: string, output: string) {
+async function removeVendor(output: string) {
   // Remove vendor directory
   try {
-    await Deno.remove(join(root, output), { recursive: true });
+    await Deno.remove(output, { recursive: true });
   } catch {
     // Do nothing
   }
 
   // Revert the import_map.json file config
-  try {
-    await Deno.stat(join(root, "import_map.json"));
-    const denoConfig = await getDenoConfig();
-    const config = denoConfig?.config || {};
+  const denoConfig = await readDenoConfig("./import_map.json");
 
-    if (config.importMap?.startsWith(output)) {
-      config.importMap = "import_map.json";
-      await Deno.writeTextFile(
-        "deno.json",
-        JSON.stringify(config, null, 2) + "\n",
-      );
-    }
-  } catch {
-    // Do nothing
+  if (denoConfig) {
+    await writeDenoConfig(denoConfig);
+  } else {
+    throw new Error("No deno.json file found");
   }
 }